improve editing experience w/ content assistance #132 #133
