+++ src/extension.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
-import * as path from 'path';
+import { isAsyncAPIFile, openAsyncAPI, openAsyncapiFiles, previewAsyncAPI } from './PreviewWebPanel';
+import { asyncapiSmartPaste } from './SmartPasteCommand';
 
-const openAsyncapiFiles: { [id: string]: vscode.WebviewPanel } = {}; // vscode.Uri.fsPath => vscode.WebviewPanel
 
 export function activate(context: vscode.ExtensionContext) {
   console.log('Congratulations, your extension "asyncapi-preview" is now active!');
@@ -30,117 +30,10 @@ export function activate(context: vscode.ExtensionContext) {
     }
   });
 
-  let disposable = vscode.commands.registerCommand('asyncapi.preview', async (uri: vscode.Uri) => {
-    uri = uri || (await promptForAsyncapiFile());
-    if (uri) {
-      console.log('Opening asyncapi file', uri.fsPath);
-      openAsyncAPI(context, uri);
-    }
-  });
 
-  context.subscriptions.push(disposable);
-}
+  context.subscriptions.push(vscode.commands.registerCommand('asyncapi.preview', previewAsyncAPI(context)));
 
-function isAsyncAPIFile(document?: vscode.TextDocument) {
-  if (!document) {
-    return false;
-  }
-  if (document.languageId === 'json') {
-    try {
-      const json = JSON.parse(document.getText());
-      return json.asyncapi;
-    } catch (e) {
-      return false;
-    }
-  }
-  if(document.languageId === 'yml' || document.languageId === 'yaml') {
-    return document.getText().match('^asyncapi:') !== null;
-  }
-  return false;
-}
-
-function openAsyncAPI(context: vscode.ExtensionContext, uri: vscode.Uri) {
-  const localResourceRoots = [
-    vscode.Uri.file(path.dirname(uri.fsPath)),
-    vscode.Uri.joinPath(context.extensionUri, 'dist/node_modules/@asyncapi/react-component/browser/standalone'),
-    vscode.Uri.joinPath(context.extensionUri, 'dist/node_modules/@asyncapi/react-component/styles'),
-  ];
-  if (vscode.workspace.workspaceFolders) {
-    vscode.workspace.workspaceFolders.forEach(folder => {
-      localResourceRoots.push(folder.uri);
-    });
-  }
-  const panel: vscode.WebviewPanel =
-    openAsyncapiFiles[uri.fsPath] ||
-    vscode.window.createWebviewPanel('asyncapi-preview', '', vscode.ViewColumn.Two, {
-      enableScripts: true,
-      retainContextWhenHidden: true,
-      localResourceRoots,
-    });
-  panel.title = path.basename(uri.fsPath);
-  panel.webview.html = getWebviewContent(context, panel.webview, uri);
-
-  panel.onDidDispose(() => {
-    delete openAsyncapiFiles[uri.fsPath];
-  });
-  openAsyncapiFiles[uri.fsPath] = panel;
-}
-
-async function promptForAsyncapiFile() {
-  if (isAsyncAPIFile(vscode.window.activeTextEditor?.document)) {
-    return vscode.window.activeTextEditor?.document.uri;
-  }
-  return await vscode.window.showOpenDialog({
-    canSelectFiles: true,
-    canSelectFolders: false,
-    canSelectMany: false,
-    openLabel: 'Open AsyncAPI file',
-    filters: {
-      AsyncAPI: ['yml', 'yaml', 'json'],
-    },
-  });
-}
-
-function getWebviewContent(context: vscode.ExtensionContext, webview: vscode.Webview, asyncapiFile: vscode.Uri) {
-  const asyncapiComponentJs = webview.asWebviewUri(
-    vscode.Uri.joinPath(context.extensionUri, 'dist/node_modules/@asyncapi/react-component/browser/standalone/index.js')
-  );
-  const asyncapiComponentCss = webview.asWebviewUri(
-    vscode.Uri.joinPath(context.extensionUri, 'dist/node_modules/@asyncapi/react-component/styles/default.min.css')
-  );
-  const asyncapiWebviewUri = webview.asWebviewUri(asyncapiFile);
-  const asyncapiBasePath = asyncapiWebviewUri.toString().replace('%2B', '+'); // this is loaded by a different library so it requires unescaping the + character
-  const html = `
-<!DOCTYPE html>
-<html>
-  <head>
-    <link rel="stylesheet" href="${asyncapiComponentCss}">
-  </head>
-  <body x-timestamp="${Date.now()}">
-    
-    <div id="asyncapi"></div>
-
-    <script src="${asyncapiComponentJs}"></script>
-    <script>
-      AsyncApiStandalone.render({
-        schema: {
-          url: '${asyncapiWebviewUri}',
-          options: { method: "GET", mode: "cors" },
-        },
-        config: {
-          show: {
-            sidebar: true,
-            errors: true,
-          },
-          parserOptions: { path: '${asyncapiBasePath}' }
-        },
-      }, document.getElementById('asyncapi'));
-    </script>
-
-  </body>
-</html>
-  `;
-  return html;
+  context.subscriptions.push(vscode.commands.registerCommand("asyncapi.paste", asyncapiSmartPaste));
 }
 
 export function deactivate() {}
